+++ app/(user-auth)/registration/page.tsx
@@ -1,12 +1,12 @@
-import DoctorRegistrationComponent from '@/components/non-user-auth/doctor-reg';
-import React from 'react'
+import DoctorRegistrationComponent from "@/components/doctor-components/doctor-reg";
+import React from "react";
 
 function page() {
   return (
     <div>
-      <DoctorRegistrationComponent/>
+      <DoctorRegistrationComponent />
     </div>
-  )
+  );
 }
 
 export default page;
