+++ app/(user-auth)/reset-password/page.tsx
@@ -1,4 +1,4 @@
-import ResetPasswordComponent from "@/components/user-auth/reset-password-component";
+import ResetPasswordComponent from "@/components/user-compoents/user-auth/reset-password-component";
 import React from "react";
 
 const ResetPassword = () => {
