+++ app/(user-protected)/(reset-section)/change-password/page.tsx
@@ -1,4 +1,4 @@
-import ChangePasswordComponent from "@/components/user-auth/change-password-component";
+import ChangePasswordComponent from "@/components/user-compoents/user-auth/change-password-component";
 import React from "react";
 
 const ChangePassword = () => {
