+++ app/admin/manage-admins/page.tsx
@@ -1,4 +1,4 @@
-import AdminReg from "@/components/non-user-auth/admin-reg-component";
+import AdminReg from "@/components/admin-components/admin-auth/admin-reg-component";
 import React from "react";
 
 const page = () => {
