+++ app/doctor/(doctor-auth)/doctor-login/page.tsx
@@ -1,4 +1,4 @@
-import DoctorLoginComponent from "@/components/doctor/doctor-login-component";
+import DoctorLoginComponent from "@/components/doctor-components/doctor-login-component";
 import React from "react";
 
 const DoctorLogin = () => {
