+++ components/doctor-components/doctor-auth/doctor-reg.tsx
