+++ lib/supabase/middleware.ts
@@ -54,9 +54,11 @@ export async function updateSession(request: NextRequest) {
     !request.nextUrl.pathname.startsWith("/signup") &&
     !request.nextUrl.pathname.startsWith("/reset-password") &&
     !request.nextUrl.pathname.startsWith("/change-password") &&
-    !request.nextUrl.pathname.startsWith("/api/")&&
-    !request.nextUrl.pathname.startsWith("/registration")&&
-    !request.nextUrl.pathname.startsWith("/admin/admin-login")
+    !request.nextUrl.pathname.startsWith("/api/") &&
+    !request.nextUrl.pathname.startsWith("/registration") &&
+    !request.nextUrl.pathname.startsWith("/admin/admin-login") &&
+    !request.nextUrl.pathname.startsWith("/doctor/doctor-login") &&
+    !request.nextUrl.pathname.startsWith("/doctor/doctor-registration")
   ) {
     // no user, potentially respond by redirecting the user to the login page
     const url = request.nextUrl.clone();
